dom/html/test/test_bug612730.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=612730
-->
<head>
  <title>Test for Bug 612730</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=612730">Mozilla Bug 612730</a>
<p id="display"></p>
<div id="content">
  <select multiple required>
    <option value="">foo</option>
    <option value="">bar</option>
  </select>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 612730 **/

SimpleTest.waitForExplicitFinish();

function runTest()
{
  var select = document.getElementsByTagName('select')[0];

  select.addEventListener("focus", function() {
    isnot(select.selectedIndex, -1, "Something should have been selected");

    ok(!select.matches(":-moz-ui-invalid"),
       ":-moz-ui-invalid should not apply");
    ok(!select.matches(":-moz-ui-valid"),
       ":-moz-ui-valid should not apply");

    SimpleTest.finish();
  }, {once: true});

  synthesizeMouse(select, 5, 5, {});
}

SimpleTest.waitForFocus(runTest);

</script>
</pre>
</body>
</html>
